Moved EnumProtocol to Sources
